backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for HandScript handwriting recognition application"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app", from = "." },
    { include = "tests", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
pillow = "^10.2.0"
transformers = "^4.36.2"
torch = { version = "^2.5.0+cu121", source = "pytorch" }
torchvision = { version = "^0.20.0+cu121", source = "pytorch" }
torchaudio = { version = "^2.5.0+cu121", source = "pytorch" }
python-docx = "^1.1.0"
reportlab = "^4.0.9"
aiofiles = "^23.2.1"
asyncpg = "^0.29.0"
httpx = "^0.27.2"
accelerate = "^1.0.1"
bitsandbytes = "^0.44.1"
peft = "^0.13.2"
pypdf = "^5.0.1"
psutil = "^6.1.0"
protobuf = "^5.28.3"
sentencepiece = "^0.2.0"
python-doctr = {extras = ["torch"], version = "^0.10.0"}
gputil = "^1.4.0"
jiwer = "^3.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.1"
isort = "^5.13.2"
httpx = "^0.27.2"
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]